requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
dj-rest-auth==5.0.2
Django==4.2.7
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.11.2
django-filter==23.4
django-nine==0.2.7
djangorestframework==3.14.0
elastic-transport==8.10.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
exceptiongroup==1.2.0
Faker==21.0.0
filelock==3.13.1
gdown==4.7.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
PySocks==1.7.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
dj-rest-auth==5.0.2
Django==4.2.7
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.11.2
django-filter==23.4
django-nine==0.2.7
djangorestframework==3.14.0
elastic-transport==8.10.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
exceptiongroup==1.2.0
Faker==21.0.0
filelock==3.13.1
gdown==4.7.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pypdf==3.17.4
PySocks==1.7.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
